% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkRDBESRawObject.R
\name{checkRDBESRawObject}
\alias{checkRDBESRawObject}
\title{Check whether an rdbesRawObject is in a valid format}
\usage{
checkRDBESRawObject(objectToCheck)
}
\arguments{
\item{objectToCheck}{rdbesRawObject i.e. a list of data.tables}
}
\value{
T if object is valid, F is object is not valid
}
\description{
Check whether an rdbesRawObject is in a valid format
}
