% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkRDBESRawObjectContent.R
\name{checkRDBESRawObjectContent}
\alias{checkRDBESRawObjectContent}
\title{checkRDBESRawObjectContent Private function to do some basic checks on
the content of the rdbesRawObject (e.g. all required field names are
present).  Function is only used by checkRDBESRawObject and should only
be passed a list of non-null objects}
\usage{
checkRDBESRawObjectContent(objectToCheck)
}
\arguments{
\item{objectToCheck}{- rdbesRawObject i.e. a list of data.tables}
}
\value{

}
\description{
checkRDBESRawObjectContent Private function to do some basic checks on
the content of the rdbesRawObject (e.g. all required field names are
present).  Function is only used by checkRDBESRawObject and should only
be passed a list of non-null objects
}
