% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineRDBESRawObjects.R
\name{combineRDBESRawObjects}
\alias{combineRDBESRawObjects}
\title{combineRDBESRawObjects combines 2 rdbesRawObjects into a single
rdbesRawObject}
\usage{
combineRDBESRawObjects(rdbesRawObject1, rdbesRawObject2)
}
\arguments{
\item{rdbesRawObject1}{The first object to combine}

\item{rdbesRawObject2}{The second object to combine}
}
\value{

}
\description{
combineRDBESRawObjects combines 2 rdbesRawObjects into a single
rdbesRawObject
}
