% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDBEPrepObj.R
\name{createDBEPrepObj}
\alias{createDBEPrepObj}
\title{Load raw object and create prepared object
Function relies on the data being
correctly named following established hierarchy}
\usage{
createDBEPrepObj(input, output)
}
\arguments{
\item{input}{a string pointing towards the input folder}

\item{output}{a string pointing towards the output folder}
}
\description{
Load raw object and create prepared object
Function relies on the data being
correctly named following established hierarchy
}
\examples{
\dontrun{
input <- "WKRDB-EST2/testData/output/DBErawObj/"
output <- "WKRDB-EST2/subGroup1/personal/John/PreparedOutputs/"

createDBEPrepObj(input = input, output = output)
}
}
