% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_token.R
\name{decode_token}
\alias{decode_token}
\title{Decode token}
\usage{
decode_token(token = rdbes_token())
}
\arguments{
\item{token}{a javascript web token got by running `rdbes_token()`}
}
\value{
list of claims
}
\description{
Decodes the token to allow inspection of claims
}
\examples{
\dontrun{
decode_token()
}

}
