% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doDBEestimantionObjUpp.R
\name{doDBEestimantionObjUpp}
\alias{doDBEestimantionObjUpp}
\title{Generates the DBE estimation object for the upper hierarchy tables}
\usage{
doDBEestimantionObjUpp(inputList)
}
\arguments{
\item{inputList}{All the data tables in a named list. Name should be equal
to the short table names e.g. DE, SD, TE, FO. An example can be found at
the share point:
https://community.ices.dk/ExpertGroups/WKRDB/2019\%20Meetings/
WKRDB-EST\%202019/06.\%20Data/Kirsten/H1/H1_upper.RData}
}
\value{
The upper hierarchy tables in the DBE estimation object
(DBEestimantionObjUpp)
}
\description{
Generates the DBE estimation object for the upper hierarchy tables
}
\examples{
\dontrun{
H1 <-
readRDS("./WKRDB-EST2/testData/output/DBErawObj/DBErawObj_DK_1966_H1.rds")
H1out <- doDBEestimantionObjUpp(H1)
}
}
