% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doDBErawObj.R
\name{doDBErawObj}
\alias{doDBErawObj}
\title{doDBErawObj
Function that reads all the csv's per table and sets up DBErawObj
based on the scripts from the repo  https://raw.githubusercontent.com/davidcurrie2001/MI_RDBES_ExchangeFiles}
\usage{
doDBErawObj(RDBESextractPath = NA, DBErawPath = NA)
}
\arguments{
\item{RDBESextractPath}{Path where the folder with the csvs exported from the RDBES are being stored. If the user does not specify it,
an interactive window will be launched, so that the user can choose the directory}

\item{DBErawPath}{Path where the function is supposed to write the DBErawObjs in. If not defined, it will be stored in the parent
of RDBESextractPath}
}
\value{
DBErawObj - the list containing all the tables that create the given hierarchy
}
\description{
doDBErawObj
Function that reads all the csv's per table and sets up DBErawObj
based on the scripts from the repo  https://raw.githubusercontent.com/davidcurrie2001/MI_RDBES_ExchangeFiles
}
