% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleFunction.R
\name{exampleFunction}
\alias{exampleFunction}
\title{Minimum Documentation Example}
\usage{
exampleFunction(stringX, stringY)
}
\arguments{
\item{stringX}{first string}

\item{stringY}{second string}
}
\value{
The  stringX and stringY pasted together
}
\description{
See the file "./tests/testthat/test-exampleFunction.R"
for defining expected input and output
}
\examples{
exampleFunction("A", "B")
}
