% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterRDBESRawObject.R
\name{filterRDBESRawObject}
\alias{filterRDBESRawObject}
\title{filterRDBESRawObject Filter an RDBESRawObject - the returned object will
include all rows which either: a) do not included any of the field names in
fieldsToFilter, or b) do include the field names and have one of the allowed
values in valuesToFilter}
\usage{
filterRDBESRawObject(rdbesRawObjectToFilter, fieldsToFilter, valuesToFilter)
}
\arguments{
\item{rdbesRawObjectToFilter}{The RDBESRawObject to filter}

\item{fieldsToFilter}{A vector of the field names you wish to check}

\item{valuesToFilter}{A vector of the field values you wish to filter for}
}
\value{

}
\description{
filterRDBESRawObject Filter an RDBESRawObject - the returned object will
include all rows which either: a) do not included any of the field names in
fieldsToFilter, or b) do include the field names and have one of the allowed
values in valuesToFilter
}
\examples{
\dontrun{
myH1RawObject <- createRDBESRawObject(rdbesExtractPath = "tests\\\\testthat\\\\h1_v_1_19")

myFields <- c("SDctry","VDctry","VDflgCtry","FTarvLoc")
myValues <- c("ZW","ZWBZH","ZWVFA" )

myFilteredObject <- filterRDBESRawObject(myH1RawObject,
                                         fieldsToFilter = myFields,
                                         valuesToFilter = myValues )
}
}
