% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateProbs.r
\name{generateProbs}
\alias{generateProbs}
\title{Generate vector of selection or inclusion probabilities
(version of fuction by John Ball)}
\usage{
generateProbs(x, probType)
}
\arguments{
\item{x}{RDBES data frame}

\item{probType}{Value == "selection" | "inclusion" ,}
}
\value{
A vector or probabilities
}
\description{
Generate vector of selection or inclusion probabilities
(version of fuction by John Ball)
}
\examples{
\dontrun{
generateProbs(x=y, probType="inclusion")
}
}
