% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTablesInHierarchies.r
\name{getRDBESTablesInHierarchies}
\alias{getRDBESTablesInHierarchies}
\title{Use the RDBES xsd files to determine which tables are
required in the different RDBES hierachies}
\usage{
getRDBESTablesInHierarchies(
  downloadFromGitHub = TRUE,
 
    gitHubFileLocation = "https://api.github.com/repos/ices-tools-dev/RDBES/contents/XSD-files",
  fileLocation
)
}
\arguments{
\item{downloadFromGitHub}{(Optional) Set to TRUE if you want to download the
lastest xsd files from GitHub}

\item{gitHubFileLocation}{(Optional) Default value is
"https://api.github.com/repos/ices-tools-dev/RDBES/contents/XSD-files"}

\item{fileLocation}{The folder to read the files from.  If you are
downloading from GitHub a copy of the latest files will be saved here.}
}
\value{
A named list with each entry being a vector of the tables required
for that hierarchy
}
\description{
Use the RDBES xsd files to determine which tables are
required in the different RDBES hierachies
}
\examples{
\dontrun{getTablesInHierarchies(downloadFromGitHub = TRUE, fileLocation = './tableDefs/')}
}
