% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello World}
\usage{
hello(name = "world")
}
\arguments{
\item{name}{a character string giving a name to greet.}
}
\value{
Invisibly returns a character string with the greeting message.
}
\description{
Print a hello world message.
}
\examples{
hello()
hello("RDBES")
}
