% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesRDBES-package.R
\docType{package}
\name{icesRDBES-package}
\alias{icesRDBES}
\alias{icesRDBES-package}
\title{Tools for the ICES Regional DataBase and Estimation System}
\description{
Tools to support the ICES Regional DataBase and Estimation System (RDBES),
to access, analyse, and process fisheries data collected under the EU Data
Collection Framework.
}
\references{
ICES Regional DataBase and Estimation System:
\url{https://sboxshiny.ices.dk/rdbescatalogue/}.
}
\author{
Colin Millar and Arni Magnusson.
}
