% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importRDBESDownloadData.R
\name{importRDBESDownloadData}
\alias{importRDBESDownloadData}
\title{Import RDBES Downloaded Data}
\usage{
importRDBESDownloadData(filenames)
}
\arguments{
\item{filenames}{- vector of paths pointing to files that should be imported}
}
\value{
a list of all the RDBES data tables
The table that are not in input data are NULL
}
\description{
Read the .zip files and/or .csv files downloadable from the ICES RDBES
web page. The function accepts a list of paths to csv and zip files. Unzips
an then uses \code{\link{createRDBESRawObject}}
}
\examples{
files <- c("./tests/testthat/h1_v_1_19/H1_2021_000_example.zip")
obj <- importRDBESDownloadData(files)
}
\seealso{
\code{\link{createRDBESRawObject}}
}
