% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapColNamesFieldR}
\alias{mapColNamesFieldR}
\title{A dataset containing the mapping from database column names
to R field names}
\format{
A data frame containing database field names and their equivalent
R field name:
\describe{
  \item{Field.Name}{The database field names}
  \item{R.Name}{The equivalent R field name}
  ...
}
}
\source{
\url{https://github.com/davidcurrie2001/MI_RDBES_ExchangeFiles}
}
\usage{
mapColNamesFieldR
}
\description{
A dataset containing the mapping from database column names
to R field names
}
\keyword{datasets}
