% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newRDBESRawObject.R
\name{newRDBESRawObject}
\alias{newRDBESRawObject}
\title{Constructor for RDBESRawObject class}
\usage{
newRDBESRawObject(
  DE = NULL,
  SD = NULL,
  VS = NULL,
  FT = NULL,
  FO = NULL,
  TE = NULL,
  LO = NULL,
  OS = NULL,
  LE = NULL,
  SS = NULL,
  SA = NULL,
  FM = NULL,
  BV = NULL,
  VD = NULL,
  SL = NULL,
  CL = NULL,
  CE = NULL
)
}
\arguments{
\item{DE}{Data table of RDBES DE data or null}

\item{SD}{Data table of RDBES DE data or null}

\item{VS}{Data table of RDBES DE data or null}

\item{FT}{Data table of RDBES DE data or null}

\item{FO}{Data table of RDBES DE data or null}

\item{TE}{Data table of RDBES DE data or null}

\item{LO}{Data table of RDBES DE data or null}

\item{OS}{Data table of RDBES DE data or null}

\item{LE}{Data table of RDBES DE data or null}

\item{SS}{Data table of RDBES DE data or null}

\item{SA}{Data table of RDBES DE data or null}

\item{FM}{Data table of RDBES DE data or null}

\item{BV}{Data table of RDBES DE data or null}

\item{VD}{Data table of RDBES DE data or null}

\item{SL}{Data table of RDBES DE data or null}

\item{CL}{Data table of RDBES DE data or null}

\item{CE}{Data table of RDBES DE data or null}
}
\value{

}
\description{
Constructor for RDBESRawObject class
}
