% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdbes_api.R
\name{rdbes_api}
\alias{rdbes_api}
\title{Get the current API URL based on the environment}
\usage{
rdbes_api(production = getOption("rdbes.production"), type = c("upload",
  "download"))
}
\arguments{
\item{production}{Logical. If TRUE, returns the production API URL. If FALSE, returns the development API URL.}

\item{type}{Character. The type of API URL to return, either "upload" or "download". Defaults to "upload".}
}
\value{
A character string representing the API URL.
}
\description{
This function returns the API URL based on whether the package is in development or production mode.
}
