% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stupidFunction.R
\name{stupidFunction}
\alias{stupidFunction}
\title{This is a stupid test function}
\usage{
stupidFunction(x, y)
}
\arguments{
\item{x}{A string}

\item{y}{Another string}
}
\value{
The strings \code{x} and \code{y} pasted together
}
\description{
This is a stupid test function
}
\examples{
stupidFunction("A","B")
}
