% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tablesInRDBESHierarchies}
\alias{tablesInRDBESHierarchies}
\title{The tables required for each RDBES hierarchy.}
\format{
A named list containing the tables required for each RDBES hierachy.
Each entry in the list is a character vecor
\describe{
  \item{tablesInRDBESHierarchies}{named list of character vectors}
  ...
}
}
\source{
\url{https://github.com/davidcurrie2001/MI_RDBES_ExchangeFiles}
}
\usage{
tablesInRDBESHierarchies
}
\description{
A list containing the tables required for each RDBES hierachy
}
\keyword{datasets}
