% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateRDBESRawObject.R
\name{validateRDBESRawObject}
\alias{validateRDBESRawObject}
\alias{checkRDBESRawObject}
\title{Check whether an rdbesRawObject is in a valid format}
\usage{
validateRDBESRawObject(objectToCheck, checkDataTypes = FALSE, verbose = TRUE)
}
\arguments{
\item{objectToCheck}{rdbesRawObject i.e. a list of data.tables}

\item{checkDataTypes}{(Optional) Set to TRUE if you want to check that
the data types of the required columns are correct, or FALSE if you don't
care.
Default value is FALSE.}

\item{verbose}{(Optional) Set to TRUE if you want informative text printed
out, or FALSE if you don't.  The default is TRUE.}
}
\value{
TRUE if object is valid, FALSE is object is not valid
}
\description{
Check whether an rdbesRawObject is in a valid format
}
