% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateRDBESRawObjectDataTypes.R
\name{validateRDBESRawObjectDataTypes}
\alias{validateRDBESRawObjectDataTypes}
\title{checkRDBESRawObjectDataTypes Checks the data types of the columns in an
rdbesRawObject against an expected list of data types.  Any differences
are returned}
\usage{
validateRDBESRawObjectDataTypes(objectToCheck)
}
\arguments{
\item{objectToCheck}{An rdbesRawObject to check}
}
\value{
A data frame containing any data type differences (an empty data
frame if there are no differences)
}
\description{
checkRDBESRawObjectDataTypes Checks the data types of the columns in an
rdbesRawObject against an expected list of data types.  Any differences
are returned
}
