% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdbes_api.R
\name{rdbes_api}
\alias{rdbes_api}
\title{Get the current API URL based on the environment}
\usage{
rdbes_api()
}
\value{
A character string representing the API URL.
}
\description{
This function returns the API URL based on whether the package is in development or production mode.
}
