% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdbes_token.R
\name{rdbes_token}
\alias{rdbes_token}
\title{Get a Token to Download RDBES Data}
\usage{
rdbes_token()
}
\value{
Character. The access token string.
}
\description{
This function authenticates with Azure using the tenant ID, client app ID, and resource specified in the package options to get an access token.
}
\examples{
\dontrun{
  token <- rdbes_token()
  # Inspect claims (out of curiosity)
  decode_token(token)
}

}
