% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdbes_upload_data.R
\name{rdbes_upload_data}
\alias{rdbes_upload_data}
\title{Upload RDBES Data}
\usage{
rdbes_upload_data(file_path, hierarchy,
  production = getOption("rdbes.production"), verbose = FALSE)
}
\arguments{
\item{file_path}{Character. The path to the file to upload.}

\item{hierarchy}{Character. The hierarchy type for the upload.}

\item{production}{Logical. Optional. Whether to use the production API endpoint. Defaults to getOption("rdbes.production").}

\item{verbose}{Logical. Optional. Whether to print verbose HTTP request/response details. Defaults to FALSE.}
}
\value{
Character. The path to the downloaded ZIP file.
}
\description{
This function authenticates with Azure, uploads a file, starts a screening job, polls for completion,
and downloads the resulting reordered file if available.
}
\examples{
\dontrun{
filename <- system.file("test_files/importtest_HNI.csv", package = "icesRDBES")

result <- rdbes_upload_data(file_path = filename, hierarchy = "HNI")
}

}
